ai/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
spacy==3.8.0
python-dateutil==2.8.2